Changed Game concept
in order to be able to fit requirements for the project, I changed the game from being a flappy bird like game to a game where the player is playing as a tank driving around shooting enemies.

For this i imported some tank assets and made some movement and shoot scripts.