Added Enemy, Pickup, UI, and improved scripts
I added an Enemy that can not yet move but it does rotate to the player and shoot.
I added a pickup that grants a stronger bullet to the player.
I added UI that displays whether you have a super bullet, how many lives you have and your score.
I improved some scripts for movement and animation.