Fixed particle shooter, changed end screen background,  fixed enemy bug
Changed the particle shooter to work more efficiently by removing particles at a certain Y position instead of after a set amount of seconds.

Added the globe with spinning tank from the main menu to the end screen.

Fixed a bug that caused the enemies to no longer aim at the player.